Move unit tests into source files. Added system tests.
