+++ src/measurement.rs
@@ -14,7 +14,7 @@
 /// }
 ///
 /// impl Measurement for Cubits {
-///     fn get_base_units(&self) -> f64 {
+///     fn as_base_units(&self) -> f64 {
 ///         self.forearms
 ///     }
 ///
@@ -22,7 +22,7 @@
 ///         Cubits { forearms: units }
 ///     }
 ///
-///    fn get_base_units_name(&self) -> &'static str {
+///    fn as_base_units_name(&self) -> &'static str {
 ///        "cu"
 ///    }
 /// }
@@ -30,7 +30,7 @@
 /// // Invoke the macro to automatically implement Add, Sub, etc...
 /// implement_measurement! { Cubits }
 ///
-/// // The main function here is only included to make doc tests compile.
+/// // The main function here is only included to make doc test_utils compile.
 /// // You should't need it in your own code.
 /// fn main() { }
 /// ```
@@ -43,7 +43,7 @@ pub trait Measurement {
     /// The default implementation always selects the base unit. Override in your
     /// Measurement  impl to select better units if required.
     fn get_appropriate_units(&self) -> (&'static str, f64) {
-        (self.get_base_units_name(), self.get_base_units())
+        (self.as_base_units_name(), self.as_base_units())
     }
 
     /// Given a list of units and their scale relative to the base unit,
@@ -53,20 +53,20 @@ pub trait Measurement {
     /// ("kilometre", 10e3)
     fn pick_appropriate_units(&self, list: &[(&'static str, f64)]) -> (&'static str, f64) {
         for &(ref unit, ref scale) in list.iter().rev() {
-            let value = self.get_base_units() / scale;
+            let value = self.as_base_units() / scale;
             if value.abs() > 1.0 {
                 return (unit, value);
             }
         }
-        (list[0].0, self.get_base_units() / list[0].1)
+        (list[0].0, self.as_base_units() / list[0].1)
     }
 
     /// Return the base unit for this type, as a string.
     /// For example "kilograms"
-    fn get_base_units_name(&self) -> &'static str;
+    fn as_base_units_name(&self) -> &'static str;
 
     /// Get this quantity in the base units
-    fn get_base_units(&self) -> f64;
+    fn as_base_units(&self) -> f64;
 
     /// Create a new quantity from the base units
     fn from_base_units(units: f64) -> Self;
@@ -102,7 +102,7 @@ macro_rules! implement_measurement {
             type Output = Self;
 
             fn add(self, rhs: Self) -> Self {
-                Self::from_base_units(self.get_base_units() + rhs.get_base_units())
+                Self::from_base_units(self.as_base_units() + rhs.as_base_units())
             }
         }
 
@@ -110,7 +110,7 @@ macro_rules! implement_measurement {
             type Output = Self;
 
             fn sub(self, rhs: Self) -> Self {
-                Self::from_base_units(self.get_base_units() - rhs.get_base_units())
+                Self::from_base_units(self.as_base_units() - rhs.as_base_units())
             }
         }
 
@@ -120,7 +120,7 @@ macro_rules! implement_measurement {
             type Output = f64;
 
             fn div(self, rhs: Self) -> f64 {
-                self.get_base_units() / rhs.get_base_units()
+                self.as_base_units() / rhs.as_base_units()
             }
         }
 
@@ -130,7 +130,7 @@ macro_rules! implement_measurement {
             type Output = Self;
 
             fn div(self, rhs: f64) -> Self {
-                Self::from_base_units(self.get_base_units() / rhs)
+                Self::from_base_units(self.as_base_units() / rhs)
             }
         }
 
@@ -140,7 +140,7 @@ macro_rules! implement_measurement {
             type Output = Self;
 
             fn mul(self, rhs: f64) -> Self {
-                Self::from_base_units(self.get_base_units() * rhs)
+                Self::from_base_units(self.as_base_units() * rhs)
             }
         }
 
@@ -156,13 +156,13 @@ macro_rules! implement_measurement {
         impl ::std::cmp::Eq for $t { }
         impl ::std::cmp::PartialEq for $t {
             fn eq(&self, other: &Self) -> bool {
-                self.get_base_units() == other.get_base_units()
+                self.as_base_units() == other.as_base_units()
             }
         }
 
         impl ::std::cmp::PartialOrd for $t {
             fn partial_cmp(&self, other: &Self) -> Option<::std::cmp::Ordering> {
-                self.get_base_units().partial_cmp(&other.get_base_units())
+                self.as_base_units().partial_cmp(&other.as_base_units())
             }
         }
     )*)
