+++ src/prelude.rs
@@ -1,0 +1,4 @@
+//! Traits that it is useful to have in scope.
+//! You should wildcard import this file.
+
+pub use super::Measurement;
