+++ src/pressure.rs
@@ -105,7 +105,7 @@ impl Pressure {
 }
 
 impl Measurement for Pressure {
-    fn get_base_units(&self) -> f64 {
+    fn as_base_units(&self) -> f64 {
         self.pascals
     }
 
@@ -113,7 +113,7 @@ impl Measurement for Pressure {
         Self::from_pascals(units)
     }
 
-    fn get_base_units_name(&self) -> &'static str {
+    fn as_base_units_name(&self) -> &'static str {
         "Pa"
     }
