+++ src/temperature.rs
@@ -122,7 +122,7 @@ impl Temperature {
 }
 
 impl Measurement for Temperature {
-    fn get_base_units(&self) -> f64 {
+    fn as_base_units(&self) -> f64 {
         self.degrees_kelvin
     }
 
@@ -130,13 +130,13 @@ impl Measurement for Temperature {
         Self::from_kelvin(degrees_kelvin)
     }
 
-    fn get_base_units_name(&self) -> &'static str {
+    fn as_base_units_name(&self) -> &'static str {
         "K"
     }
 }
 
 impl Measurement for TemperatureDelta {
-    fn get_base_units(&self) -> f64 {
+    fn as_base_units(&self) -> f64 {
         self.kelvin_degrees
     }
 
@@ -144,7 +144,7 @@ impl Measurement for TemperatureDelta {
         Self::from_kelvin(kelvin_degrees)
     }
 
-    fn get_base_units_name(&self) -> &'static str {
+    fn as_base_units_name(&self) -> &'static str {
         "K"
     }
 }
@@ -184,15 +184,133 @@ impl ::std::ops::Sub<Temperature> for Temperature {
 impl ::std::cmp::Eq for Temperature {}
 impl ::std::cmp::PartialEq for Temperature {
     fn eq(&self, other: &Self) -> bool {
-        self.get_base_units() == other.get_base_units()
+        self.as_base_units() == other.as_base_units()
     }
 }
 
 impl ::std::cmp::PartialOrd for Temperature {
     fn partial_cmp(&self, other: &Self) -> Option<::std::cmp::Ordering> {
-        self.get_base_units().partial_cmp(&other.get_base_units())
+        self.as_base_units().partial_cmp(&other.as_base_units())
     }
 }
 
 implement_display!(Temperature);
 implement_measurement!(TemperatureDelta);
+
+#[cfg(test)]
+mod test {
+    use temperature::*;
+    use test_utils::assert_almost_eq;
+
+    // Temperature Units
+    #[test]
+    fn kelvin() {
+        let t = Temperature::from_kelvin(100.0);
+        let o = t.as_kelvin();
+
+        assert_almost_eq(o, 100.0);
+    }
+
+    #[test]
+    fn celsius() {
+        let t = Temperature::from_kelvin(100.0);
+        let o = t.as_celsius();
+
+        assert_almost_eq(o, -173.15);
+    }
+
+    #[test]
+    fn fahrenheit() {
+        let t = Temperature::from_kelvin(100.0);
+        let o = t.as_fahrenheit();
+
+        assert_almost_eq(o, -279.67);
+    }
+
+    #[test]
+    fn rankine() {
+        let t = Temperature::from_kelvin(100.0);
+        let o = t.as_rankine();
+
+        assert_almost_eq(o, 180.0);
+    }
+
+    // Traits
+    #[test]
+    fn add() {
+        let a = Temperature::from_kelvin(2.0);
+        let b = TemperatureDelta::from_kelvin(4.0);
+        let c = a + b;
+        let d = b + a;
+        assert_almost_eq(c.as_kelvin(), 6.0);
+        assert_eq!(c, d);
+    }
+
+    #[test]
+    fn add2() {
+        let a = TemperatureDelta::from_kelvin(2.0);
+        let b = TemperatureDelta::from_kelvin(4.0);
+        let c = a + b;
+        let d = b + a;
+        assert_almost_eq(c.as_kelvin(), 6.0);
+        assert_eq!(c, d);
+    }
+
+    #[test]
+    fn sub() {
+        let a = Temperature::from_kelvin(4.0);
+        let b = TemperatureDelta::from_kelvin(2.0);
+        let c = a - b;
+        assert_almost_eq(c.as_kelvin(), 2.0);
+    }
+
+    #[test]
+    fn sub2() {
+        let a = Temperature::from_fahrenheit(212.0);
+        let b = Temperature::from_celsius(75.0);
+        let c = a - b;
+        assert_almost_eq(c.as_kelvin(), 25.0);
+    }
+
+    #[test]
+    fn sub3() {
+        let a = TemperatureDelta::from_fahrenheit(180.0);
+        let b = TemperatureDelta::from_celsius(75.0);
+        let c = a - b;
+        assert_almost_eq(c.as_kelvin(), 25.0);
+    }
+
+    #[test]
+    fn mul() {
+        let a = TemperatureDelta::from_celsius(5.0);
+        let b = a * 2.0;
+        let c = 2.0 * a;
+        assert_almost_eq(b.as_celsius(), 10.0);
+        assert_eq!(b, c);
+    }
+
+    #[test]
+    fn eq() {
+        let a = Temperature::from_kelvin(2.0);
+        let b = Temperature::from_kelvin(2.0);
+        assert_eq!(a == b, true);
+    }
+
+    #[test]
+    fn neq() {
+        let a = Temperature::from_kelvin(2.0);
+        let b = Temperature::from_kelvin(4.0);
+        assert_eq!(a == b, false);
+    }
+
+    #[test]
+    fn cmp() {
+        let a = Temperature::from_kelvin(2.0);
+        let b = Temperature::from_kelvin(4.0);
+        assert_eq!(a < b, true);
+        assert_eq!(a <= b, true);
+        assert_eq!(a > b, false);
+        assert_eq!(a >= b, false);
+    }
+
+}
